Refactor: remove stupid message. Fix pigeon signal nullity
